Update AUIPlugin to 5.9.17
When connecting other Atlassian Applications with Crowd the process fails, because of crowds outdated auiplugin. This patch bumps to auiplugin 5.9.17, which is new enough for the application-links-application and old enough for the rest of crowd included assets. Maybe someday ATL will update this plugin and the added 4 lines can be removed.